Ignore obsolete agent_nodes directory
